delete IoTDBSqlVisitor
